Rearrange navbar and complete flow
