+++ frontend/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next'
 import { Roboto_Mono } from 'next/font/google'
 import './globals.css'
-import NavLinks from '../components/navbar'
+import Navbar from '../components/Navbar'
 
 const roboto = Roboto_Mono({ subsets: ['latin'], weight: ['400', '700'] })
 
@@ -19,7 +19,7 @@ export default function RootLayout({
     <html lang='en'>
       <body className={roboto.className} style={{ backgroundColor: '#323437' }}>
         <div className='flex flex-col min-h-screen'>
-          <NavLinks />
+          <Navbar />
           <main className='flex-1 flex items-center justify-center'>{children}</main>
         </div>
       </body>
