+++ frontend/src/components/Navbar/index.tsx
@@ -1,0 +1,67 @@
+import Link from 'next/link'
+import { getSession } from '@/auth'
+import NavbarLinks from './NavbarLinks'
+import LogoutButton from '../LogoutButton'
+
+export default async function Navbar() {
+  const session = await getSession()
+  const club = session?.club
+
+  return (
+    <>
+      <header className='flex items-center w-full justify-between p-4 border-b border-gray-700'>
+        <div className='flex items-center w-full gap-10'>
+          <div className='flex items-center space-x-2'>
+            <Link href='/'>
+              <div className='flex items-center space-x-5'>
+                <RabbitIcon className='h-6 w-6' />
+                <h1 className='text-xl font-bold'>ClubsConnect</h1>
+              </div>
+            </Link>
+          </div>
+          <NavbarLinks />
+        </div>
+        <div className='flex gap-10 items-center'>
+          {club ? (
+            <>
+              {club.name}
+              <LogoutButton />
+            </>
+          ) : (
+            <Link
+              href='/login'
+              className={
+                'text-nowrap inline-flex h-10 items-center justify-center rounded-md border border-input bg-background px-8 text-sm font-medium shadow-sm transition-colors hover:bg-accent hover:text-accent-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50'
+              }
+            >
+              Login as a Club
+            </Link>
+          )}
+        </div>
+      </header>
+    </>
+  )
+}
+
+function RabbitIcon(props: React.SVGProps<SVGSVGElement>) {
+  return (
+    <svg
+      {...props}
+      xmlns='http://www.w3.org/2000/svg'
+      width='24'
+      height='24'
+      viewBox='0 0 24 24'
+      fill='none'
+      stroke='currentColor'
+      strokeWidth='2'
+      strokeLinecap='round'
+      strokeLinejoin='round'
+    >
+      <path d='M13 16a3 3 0 0 1 2.24 5' />
+      <path d='M18 12h.01' />
+      <path d='M18 21h-8a4 4 0 0 1-4-4 7 7 0 0 1 7-7h.2L9.6 6.4a1 1 0 1 1 2.8-2.8L15.8 7h.2c3.3 0 6 2.7 6 6v1a2 2 0 0 1-2 2h-1a3 3 0 0 0-3 3' />
+      <path d='M20 8.54V4a2 2 0 1 0-4 0v3' />
+      <path d='M7.612 12.524a3 3 0 1 0-1.6 4.3' />
+    </svg>
+  )
+}
